pyproject.toml
[project]
name = "quorus"
version = "0.4.0"
description = "Real-time coordination layer for AI agent swarms — rooms, messaging, distributed locks"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Arav Kekane", email = "aravkekane@gmail.com"},
    {name = "Aarya"},
]
readme = "README.md"
keywords = ["ai", "agents", "coordination", "mcp", "multi-agent", "swarm", "relay"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "uvicorn>=0.29.0",
    "mcp>=1.2.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "redis[hiredis]>=5.0.0",
    "anthropic>=0.94.0",
]

[project.urls]
Homepage = "https://quorus.dev"
Repository = "https://github.com/Aarya2004/quorus"
Issues = "https://github.com/Aarya2004/quorus/issues"

[project.scripts]
quorus = "quorus_cli.cli:main"
quorus-relay = "quorus.relay:main"
quorus-analytics = "quorus.analytics:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
# Dashboard contains inline HTML/CSS/JS in Python strings
per-file-ignores = {"quorus/dashboard.py" = ["E501"]}
select = ["E", "F", "I", "W"]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "testcontainers[redis]>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "error::RuntimeWarning",
    "ignore::jwt.warnings.InsecureKeyLengthWarning",
]
markers = [
    "integration: marks tests as integration tests requiring Docker (deselect with '-m \"not integration\"')",
]

[tool.hatch.build.targets.wheel]
packages = ["quorus", "packages/sdk/quorus_sdk", "packages/tui/quorus_tui", "packages/mcp/quorus_mcp", "packages/cli/quorus_cli"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.quorus",
    "*.db",
    "*.db-shm",
    "*.db-wal",
    "/website",
    "/tests",
    "/dist",
    "/.venv",
    "/.github",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
